chore(execd): extract safego to internal common package
